6: extracted components to different files
