+++ core/uicomponents/src/main/java/com/mateuszholik/uicomponents/extensions/LocalDateExtensions.kt
@@ -3,7 +3,7 @@ package com.mateuszholik.uicomponents.extensions
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-internal fun LocalDate.asString(): String {
+internal fun LocalDate.asHourString(): String {
 
     val dateTimeFormatter = DateTimeFormatter.ofPattern("dd.MM")
