+++ core/uicomponents/src/main/java/com/mateuszholik/uicomponents/extensions/LocalDateTimeExtensions.kt
@@ -3,7 +3,7 @@ package com.mateuszholik.uicomponents.extensions
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
-internal fun LocalDateTime.asString(): String {
+internal fun LocalDateTime.asHourString(): String {
     val dateTimeFormatter = DateTimeFormatter.ofPattern("HH:mm")
 
     return this.format(dateTimeFormatter)
