+++ core/uicomponents/src/main/java/com/mateuszholik/uicomponents/team/HorizontalTeamItem.kt
@@ -1,38 +1,26 @@
-package com.mateuszholik.uicomponents.match
+package com.mateuszholik.uicomponents.team
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import com.mateuszholik.designsystem.R
+import com.mateuszholik.designsystem.theme.FootballScoreTheme
 import com.mateuszholik.designsystem.theme.sizing
 import com.mateuszholik.designsystem.theme.spacing
 import com.mateuszholik.designsystem.theme.textSizing
-import com.mateuszholik.model.Score
 import com.mateuszholik.model.Team
-import com.mateuszholik.designsystem.R
 import com.mateuszholik.uicomponents.images.Image
 
 @Composable
-internal fun Teams(
-    modifier: Modifier,
-    homeTeam: Team,
-    awayTeam: Team,
-) {
-    Column(modifier = modifier) {
-        TeamItem(team = homeTeam)
-        TeamItem(team = awayTeam)
-    }
-}
-
-@Composable
-internal fun TeamItem(team: Team) {
+internal fun HorizontalTeamItem(team: Team) {
     Row(
         modifier = Modifier.padding(
             top = MaterialTheme.spacing.tiny,
@@ -57,32 +45,20 @@ internal fun TeamItem(team: Team) {
     }
 }
 
+@Preview
 @Composable
-internal fun Score(
-    modifier: Modifier,
-    score: Score,
-    scoreColor: Color = MaterialTheme.colorScheme.onSurface,
-) {
-    Column(modifier = modifier) {
-        Text(
-            modifier = Modifier.padding(
-                vertical = MaterialTheme.spacing.tiny,
-                horizontal = MaterialTheme.spacing.extraSmall,
-            ),
-            text = "${score.home}",
-            fontSize = MaterialTheme.textSizing.small,
-            fontWeight = FontWeight.Bold,
-            color = scoreColor
-        )
-        Text(
-            modifier = Modifier.padding(
-                vertical = MaterialTheme.spacing.tiny,
-                horizontal = MaterialTheme.spacing.extraSmall,
-            ),
-            text = "${score.away}",
-            fontSize = MaterialTheme.textSizing.small,
-            fontWeight = FontWeight.Bold,
-            color = scoreColor
-        )
+private fun Preview() {
+    FootballScoreTheme {
+        Surface(color = MaterialTheme.colorScheme.surface) {
+            HorizontalTeamItem(
+                team = Team(
+                    crest = "",
+                    id = 1,
+                    name = "Manchester United",
+                    shortName = "ManU",
+                    tla = "MU"
+                )
+            )
+        }
     }
 }
